Refactor code for better consistency and readability:
- Simplified type imports by removing redundant `std` prefixes.
- Replaced manual error propagation with the `?` operator for cleaner error handling.
- Transitioned from `std::thread::sleep` to `tokio::time::sleep` for consistency with async context.
- Removed redundant `as` casting in favor of direct type literals (e.g., `0f64`) in comparisons.
- Cleaned up unused `TypeId` qualifiers by directly referencing type constants.
- Optimized comparison logic and redundant `unwrap` calls with error handling improvements.
- Enhanced method logic by removing unnecessary return statements and redundant unwraps.

+++ src/types_db/bigint_type.rs
@@ -21,7 +21,7 @@ impl Type for BigIntType {
             Val::Integer(r) => CmpBool::from(0 == *r as i64),
             Val::SmallInt(r) => CmpBool::from(0 == *r as i64),
             Val::TinyInt(r) => CmpBool::from(0 == *r as i64),
-            Val::Decimal(r) => CmpBool::from((0 as f64) == *r),
+            Val::Decimal(r) => CmpBool::from((0f64) == *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
@@ -41,7 +41,7 @@ impl Type for BigIntType {
             Val::Integer(r) => CmpBool::from(0 < *r as i64),
             Val::SmallInt(r) => CmpBool::from(0 < *r as i64),
             Val::TinyInt(r) => CmpBool::from(0 < *r as i64),
-            Val::Decimal(r) => CmpBool::from((0 as f64) < *r),
+            Val::Decimal(r) => CmpBool::from((0f64) < *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
@@ -53,7 +53,7 @@ impl Type for BigIntType {
             Val::Integer(r) => CmpBool::from(0 <= *r as i64),
             Val::SmallInt(r) => CmpBool::from(0 <= *r as i64),
             Val::TinyInt(r) => CmpBool::from(0 <= *r as i64),
-            Val::Decimal(r) => CmpBool::from((0 as f64) <= *r),
+            Val::Decimal(r) => CmpBool::from((0f64) <= *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
@@ -65,7 +65,7 @@ impl Type for BigIntType {
             Val::Integer(r) => CmpBool::from(0 > *r as i64),
             Val::SmallInt(r) => CmpBool::from(0 > *r as i64),
             Val::TinyInt(r) => CmpBool::from(0 > *r as i64),
-            Val::Decimal(r) => CmpBool::from((0 as f64) > *r),
+            Val::Decimal(r) => CmpBool::from((0f64) > *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
@@ -77,7 +77,7 @@ impl Type for BigIntType {
             Val::Integer(r) => CmpBool::from(0 >= *r as i64),
             Val::SmallInt(r) => CmpBool::from(0 >= *r as i64),
             Val::TinyInt(r) => CmpBool::from(0 >= *r as i64),
-            Val::Decimal(r) => CmpBool::from((0 as f64) >= *r),
+            Val::Decimal(r) => CmpBool::from((0f64) >= *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
