+++ src/types_db/integer_type.rs
@@ -29,7 +29,7 @@ impl Type for IntegerType {
                     CmpBool::CmpFalse
                 }
             }
-            Val::Decimal(r) => CmpBool::from((0 as f64) == *r),
+            Val::Decimal(r) => CmpBool::from((0f64) == *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
@@ -55,7 +55,7 @@ impl Type for IntegerType {
                     CmpBool::from(*r > 0)
                 }
             }
-            Val::Decimal(r) => CmpBool::from((0 as f64) < *r),
+            Val::Decimal(r) => CmpBool::from((0f64) < *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
@@ -73,7 +73,7 @@ impl Type for IntegerType {
                     CmpBool::from(*r > 0)
                 }
             }
-            Val::Decimal(r) => CmpBool::from((0 as f64) <= *r),
+            Val::Decimal(r) => CmpBool::from((0f64) <= *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
@@ -91,7 +91,7 @@ impl Type for IntegerType {
                     CmpBool::from(*r < 0)
                 }
             }
-            Val::Decimal(r) => CmpBool::from((0 as f64) > *r),
+            Val::Decimal(r) => CmpBool::from((0f64) > *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
@@ -109,7 +109,7 @@ impl Type for IntegerType {
                     CmpBool::from(*r < 0)
                 }
             }
-            Val::Decimal(r) => CmpBool::from((0 as f64) >= *r),
+            Val::Decimal(r) => CmpBool::from((0f64) >= *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
