+++ src/types_db/smallint_type.rs
@@ -34,7 +34,7 @@ impl Type for SmallIntType {
                     CmpBool::CmpFalse
                 }
             }
-            Val::Decimal(r) => CmpBool::from((0 as f64) == *r),
+            Val::Decimal(r) => CmpBool::from((0f64) == *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
@@ -66,7 +66,7 @@ impl Type for SmallIntType {
                     CmpBool::from(*r > 0)
                 }
             }
-            Val::Decimal(r) => CmpBool::from((0 as f64) < *r),
+            Val::Decimal(r) => CmpBool::from((0f64) < *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
@@ -90,7 +90,7 @@ impl Type for SmallIntType {
                     CmpBool::from(*r > 0)
                 }
             }
-            Val::Decimal(r) => CmpBool::from((0 as f64) <= *r),
+            Val::Decimal(r) => CmpBool::from((0f64) <= *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
@@ -114,7 +114,7 @@ impl Type for SmallIntType {
                     CmpBool::from(*r < 0)
                 }
             }
-            Val::Decimal(r) => CmpBool::from((0 as f64) > *r),
+            Val::Decimal(r) => CmpBool::from((0f64) > *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
@@ -138,7 +138,7 @@ impl Type for SmallIntType {
                     CmpBool::from(*r < 0)
                 }
             }
-            Val::Decimal(r) => CmpBool::from((0 as f64) >= *r),
+            Val::Decimal(r) => CmpBool::from((0f64) >= *r),
             Val::Null => CmpBool::CmpNull,
             _ => CmpBool::CmpFalse,
         }
